% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aadt_data.R
\name{add_aadt}
\alias{add_aadt}
\title{add average annual daily traffic to geocoded data}
\usage{
add_aadt(d, buffer_radius = 400, ...)
}
\arguments{
\item{d}{data.frame or tibble with columns called 'lat', 'lon'}

\item{buffer_radius}{buffer radius in meters, defaults to 400 m}

\item{...}{arguments passed to \code{s3::s3_get_files()}}
}
\value{
the input dataframe, with lengths, vehicle meters for all vehicle
types, and vehicle meters for trucks for both moving (interstates, freeways,
and expressways) and stop-and-go (arterial roads) traffic.
}
\description{
add average annual daily traffic to geocoded data
}
\examples{
if (FALSE) {
d <- tibble::tribble(
     ~id,         ~lat,    ~lon,
     '55000100280', 39.2, -84.6,
     '55000100281', 39.2, -84.6,
     '55000100282', 39.2, -84.6)

   add_aadt(d)
}
}
